chore: Move `<Seekbar />` to `/current-track` screen
- Since this is the only place where we'll use it; in addition, we'll have the imports ready when implementing the volume slider
